requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
exceptiongroup==1.2.2
faiss-cpu==1.10.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
google-auth==2.38.0
google-genai==1.8.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.7
langsmith==0.3.19
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.2.0
multitasking==0.0.11
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
orjson==3.10.16
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==4.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.18.3
yfinance==0.2.55
zstandard==0.23.0
